requirements.txt
"""
Requirements file for the Tradeovate trading bot.
"""
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
pyyaml>=6.0
websocket-client>=1.2.0
requests>=2.26.0
joblib>=1.1.0

# API integration
ccxt>=1.60.0

# Logging and utilities
tqdm>=4.62.0
colorlog>=6.6.0
